refactoring: move sanitize and toObject code out of schema file
